Domain updates:     Income.cs: Changed ID type from byte to int. EntityFramework doesn't make identity fields for Byte ID fields.     Language.cs: Renamed from EnglishLevel.cs; byte->int for ID     Neighborhood.cs: rename & byte -> int     Person.cs: Added Createdby & Updatedby fields     Race.cs: rename & byte -> int     Skill.cs: rename & byte -> int, added 'skill'field     Worker.cs: changed RaceID name

+++ Machete.Domain/Entities/Income.cs
@@ -10,7 +10,7 @@ namespace Machete.Domain
 {
     public class Income
     {
-        public byte ID { get; set; }
+        public int ID { get; set; }
         public string incomelabel { get; set; }
     }
 }
